streamlit/requirements.txt
streamlit
pandas
sqlalchemy
psycopg2-binary
altair
matplotlib